third_party/blink/web_tests/external/wpt/css/css-text/text-autospace/text-autospace-001-expected.txt

This is a testharness.js-based test.
FAIL text-autospace-001 assert_array_equals: expected property 1 to be 45 but got 40 (expected array [40, 45, 40, 45, 40] got [40, 40, 40, 40, 40])
FAIL text-autospace-001 1 assert_array_equals: expected property 3 to be 45 but got 40 (expected array [40, 40, 40, 45, 40] got [40, 40, 40, 40, 40])
FAIL text-autospace-001 2 assert_array_equals: expected property 1 to be 90 but got 80 (expected array [80, 90, 80, 90, 80, 40, 45, 40, 45, 40] got [80, 80, 80, 80, 80, 40, 40, 40, 40, 40])
Harness: the test ran to completion.
